net/spv: move rest interface to it's own respective file
